Make the sandbox UI provider-aware
Add a cross-provider availability model (projectStore.availableSandboxProviders
from sandbox:status) and nono config IPC (profile, network toggle, profile
list). Split SandboxSection into a provider router over the verbatim Lima
section and a new nono section. Gate settings on any backend; fix Lima-only
copy. The kanban card and terminal header offer one 'Open in <backend>
sandbox' entry per installed backend (which also persists the task's backend
so runners/hooks follow) plus a clear option. Status dots and card labels
show the backend name.
