+++ src/__tests__/renderer/setup.ts
@@ -127,6 +127,9 @@ const mockApi = {
   },
   sandbox: {
     status: vi.fn().mockResolvedValue([]),
+    nonoConfig: vi.fn().mockResolvedValue({}),
+    setNonoConfig: vi.fn().mockResolvedValue({ success: true }),
+    nonoProfiles: vi.fn().mockResolvedValue([]),
   },
   globalSettings: {
     get: vi.fn().mockResolvedValue(undefined),
