+++ src/components/terminal/StatusDot.tsx
@@ -1,10 +1,19 @@
 import { Tooltip } from '../ui/Tooltip';
+import type { SandboxProviderId } from '../../types';
 
 interface StatusDotProps {
   summaryType: string;
   sandboxed?: boolean;
+  /** Backend name shown in the tooltip; falls back to "Sandboxed". */
+  sandboxProvider?: SandboxProviderId;
   size?: number;
 }
+
+const PROVIDER_LABELS: Record<SandboxProviderId, string> = {
+  none: 'Sandboxed',
+  lima: 'Lima VM',
+  nono: 'nono',
+};
 
 const COLORS: Record<string, string> = {
   thinking: '#da77f2',
@@ -19,11 +28,13 @@ const LABELS: Record<string, string> = {
   error: 'Failed',
 };
 
-export function StatusDot({ summaryType, sandboxed = false, size = 6 }: StatusDotProps) {
+export function StatusDot({ summaryType, sandboxed = false, sandboxProvider, size = 6 }: StatusDotProps) {
   const isThinking = summaryType === 'thinking';
   const background = COLORS[summaryType] ?? COLORS.ready;
   const label = LABELS[summaryType] ?? LABELS.ready;
-  const tooltipText = sandboxed ? `${label} · Sandboxed` : label;
+  const sandboxLabel =
+    sandboxProvider && sandboxProvider !== 'none' ? PROVIDER_LABELS[sandboxProvider] : PROVIDER_LABELS.none;
+  const tooltipText = sandboxed ? `${label} · ${sandboxLabel}` : label;
   return (
     <Tooltip text={tooltipText} placement="top" delay={300} offsetPx={sandboxed ? 8 : 6}>
       <span
