+++ src/sandbox/nono/config.ts
@@ -1,6 +1,9 @@
+import * as os from 'node:os';
+import * as path from 'node:path';
+import * as fs from 'node:fs/promises';
 import { getGlobalSetting, setGlobalSetting } from '../../db';
 import { getLogger } from '../../logger';
-import type { NonoConfig } from './argv';
+import type { NonoConfig } from '../types';
 
 const nonoLog = getLogger().scope('nono');
 
@@ -34,4 +37,22 @@ export async function getNonoConfig(projectPath: string): Promise<NonoConfig> {
 
 export async function setNonoConfig(projectPath: string, config: NonoConfig): Promise<{ success: boolean }> {
   return setGlobalSetting(nonoConfigKey(projectPath), JSON.stringify(config));
+}
+
+/**
+ * List named nono profiles the user has installed. nono resolves `--profile`
+ * names against `~/.config/nono/profiles`; we surface them so the settings UI
+ * can offer a picker. Returns an empty list when the dir is absent.
+ */
+export async function listNonoProfiles(): Promise<string[]> {
+  const dir = path.join(os.homedir(), '.config', 'nono', 'profiles');
+  try {
+    const entries = await fs.readdir(dir, { withFileTypes: true });
+    return entries
+      .filter((e) => e.isFile() && /\.(json|toml|ya?ml)$/i.test(e.name))
+      .map((e) => e.name.replace(/\.(json|toml|ya?ml)$/i, ''))
+      .sort();
+  } catch {
+    return [];
+  }
 }
